传感器代码修改增加频率统计

+++ alarmclock/src/main/java/com/showdy/alarm/fragment/MedClockFragment.kt
@@ -9,13 +9,16 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.showdy.alarm.R
 import com.showdy.alarm.WatchViewModel
 import com.showdy.alarm.activity.MedClockActivity
+import com.showdy.alarm.service.WatchService
 import com.showdy.alarm.util.ActivityManager
 import kotlinx.android.synthetic.main.fragment_manual_clcok.*
 import kotlinx.coroutines.Dispatchers
@@ -35,6 +38,18 @@ class MedClockFragment : Fragment() {
 
     private val viewModel by viewModels<WatchViewModel>()
 
+    /**
+     * 创建的时机必须要在Fragments CREATED之前
+     */
+    private val launcher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            //回調
+            if (it.resultCode == Activity.RESULT_OK) {
+                val data = it.data?.getStringExtra(MedClockActivity.KEY_LAST_MED)
+                lastClock.text = data?: "还没服药"
+            }
+        }
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -43,6 +58,7 @@ class MedClockFragment : Fragment() {
         return inflater.inflate(R.layout.fragment_manual_clcok, container, false)
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         view.findViewById<TextView>(R.id.medClock).setOnClickListener {
@@ -61,38 +77,49 @@ class MedClockFragment : Fragment() {
 
 
         lastClock.setOnLongClickListener {
-            lifecycleScope.launch {
-                val delete = deleteFile(requireContext().applicationContext.filesDir)
-                withContext(Dispatchers.Main) {
-                    Toast.makeText(
-                        requireContext(),
-                        if (delete) "删除成功" else "删除失败",
-                        Toast.LENGTH_SHORT
-                    ).show()
+            AlertDialog.Builder(requireContext())
+                .setTitle("退出APP")
+                .setMessage("退出服务或者删除文件")
+                .setPositiveButton("删除文件") { dialog, _ ->
+                    dialog.dismiss()
+                    lifecycleScope.launch {
+                        val delete = withContext(Dispatchers.IO) {
+                            deleteFile(requireContext().applicationContext.filesDir)
+                        }
+                        withContext(Dispatchers.Main) {
+                            Toast.makeText(
+                                requireContext(),
+                                if (delete) "删除成功" else "删除失败",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                            delay(1000)
+                            //停止服务
+                            WatchService.stopService()
+                            //退出app
+                            ActivityManager.getInstance().exitApp()
 
-                    delay(1000)
-
+                        }
+                    }
+                }
+                .setNegativeButton("停止服务") { dialog, _ ->
+                    dialog.dismiss()
+                    WatchService.stopService()
+                    //退出app
                     ActivityManager.getInstance().exitApp()
                 }
-            }
+                .show()
             return@setOnLongClickListener true
         }
     }
 
 
-    override fun onResume() {
-        super.onResume()
-
-    }
-
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == MedClockActivity.REQUEST_CODE && resultCode == Activity.RESULT_OK) {
-            val time = data?.getStringExtra(MedClockActivity.KEY_LAST_MED)
-            lastClock.text = time ?: "还未服药"
-        }
-    }
+//    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+//        super.onActivityResult(requestCode, resultCode, data)
+//        if (requestCode == MedClockActivity.REQUEST_CODE && resultCode == Activity.RESULT_OK) {
+//            val time = data?.getStringExtra(MedClockActivity.KEY_LAST_MED)
+//            lastClock.text = time ?: "还未服药"
+//        }
+//    }
 
 
     /**
@@ -103,14 +130,20 @@ class MedClockFragment : Fragment() {
      */
     private fun startActivityForResult() {
         //跳转
-        Intent(requireContext(), MedClockActivity::class.java).apply {
-            startActivityForResult(this, MedClockActivity.REQUEST_CODE)
-        }
+//        Intent(requireContext(), MedClockActivity::class.java).apply {
+//            startActivityForResult(this, MedClockActivity.REQUEST_CODE)
+//        }
+
+        /**
+         * 不需要传递数据
+         */
+        launcher.launch(Intent(requireContext(), MedClockActivity::class.java))
+
     }
 
 
     @RequiresApi(Build.VERSION_CODES.O)
-    private suspend fun deleteFile(file: File): Boolean {
+    private fun deleteFile(file: File): Boolean {
         val files = file.listFiles() ?: return false
         files.indices.forEachIndexed { index, it ->
             if (files[index].isDirectory) {
